+++ src/Installer/InstallerFactory.php
@@ -57,14 +57,9 @@ class InstallerFactory
     public $dataDir;
 
     /**
-     * @var array
+     * @var ConfigDumper
      */
-    public $notPaths = [];
-
-    /**
-     * @var array
-     */
-    public $notNames = [];
+    public $dumper;
 
     /**
      * @var string
@@ -79,11 +74,7 @@ class InstallerFactory
     public function addInstallers(InstallerCollection $installers)
     {
         $installers->add(new MoodleInstaller($this->execute, $this->database, $this->moodle, new MoodleConfig(), $this->branch, $this->dataDir));
-        $installers->add(new PluginInstaller($this->moodle, $this->plugin, $this->notPaths, $this->notNames));
-
-        if (!empty($this->pluginsDir)) {
-            $installers->add(new ExtraPluginsInstaller($this->moodle, $this->pluginsDir));
-        }
+        $installers->add(new PluginInstaller($this->moodle, $this->plugin, $this->pluginsDir, $this->dumper));
         $installers->add(new VendorInstaller($this->moodle, $this->plugin, $this->execute));
 
         if ($this->plugin->hasBehatFeatures() || $this->plugin->hasUnitTests()) {